requirements.txt
appdirs==1.4.3
atomicwrites==1.1.5
attrs==19.3.0
beautifulsoup4==4.4.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
cov-core==1.15.0
coverage==4.5.2
crayons==0.4.0
cryptography==2.8
dbus-next==0.2.3
distlib==0.3.0
distro==1.4.0
filelock==3.0.12
html-table-parser-python3==0.2.0
idna==2.8
importlib-metadata==1.5.0
Jinja2>=3.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
more-itertools==4.2.0
MouseInfo==0.1.3
netifaces==0.10.4
nose2==0.9.1
numpy==1.21.1
oauthlib==3.1.0
packaging==20.3
pandas==1.3.1
pexpect==4.6.0
pluggy==0.13.0
py==1.8.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.15.5
PyScreeze==0.1.27
pyserial==3.4
pytest==4.6.9
pytest-cov==2.8.1
python-dateutil==2.8.2
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.1
PyYAML==5.3.1
requests==2.22.0
selenium==3.141.0
six==1.14.0
ssh-import-id==5.10
urllib3==1.25.8
virtualenv==20.0.17
wcwidth==0.1.8
webdriver-manager==3.4.2
zipp==1.0.0
flask==2.0.2
itsdangerous>=2.0
click>=7.1.2
Werkzeug>=2.0
MarkupSafe>=2.0
line-bot-sdk==1.0
